fix: add pooling

+++ src/modules/stream/types.js
@@ -5,7 +5,12 @@ import { getThreads, metadataManager } from "../sub/utils.js";
 import { request } from "undici";
 import { create as contentDisposition } from "content-disposition-header";
 import { AbortController } from "abort-controller"
-import { createReadStream, unlink } from "fs";
+import { access, constants, createReadStream, unlink } from "fs";
+import { createHash, randomBytes } from "crypto";
+import path from "path";
+
+const ongoingProcesses = new Map();
+const ongoingSpawn = new Map();
 
 function closeRequest(controller) {
   try { controller.abort() } catch {}
@@ -36,6 +41,33 @@ function pipe(from, to, done) {
   from.pipe(to);
 }
 
+export async function downloadVideo(info, res) {
+  const abortController = new AbortController();
+  const shutdown = () => {
+    console.log('response close, and done')
+    closeRequest(abortController);
+    closeResponse(res);
+  };
+
+  try {
+    res.setHeader('Content-disposition', contentDisposition(info.filename));
+
+    const { body: stream, headers } = await request(info.url, {
+      headers: { 'user-agent': genericUserAgent },
+      signal: abortController.signal,
+      maxRedirections: 16
+    });
+
+    res.setHeader('Connection', 'keep-alive');
+    res.setHeader('content-type', headers['content-type']);
+    res.setHeader('content-length', headers['content-length']);
+
+    pipe(stream, res, shutdown);
+  } catch {
+    shutdown();
+  }
+}
+
 export async function streamDefault(streamInfo, res) {
   const abortController = new AbortController();
   const shutdown = () => (closeRequest(abortController), closeResponse(res));
@@ -59,31 +91,125 @@ export async function streamDefault(streamInfo, res) {
   }
 }
 
-const simpleHash = str => {
-  let hash = 0;
-  for (let i = 0; i < str.length; i++) {
-    const char = str.charCodeAt(i);
-    hash = (hash << 5) - hash + char;
-    hash &= hash; // Convert to 32bit integer
-  }
-  return new Uint32Array([hash])[0].toString(36);
-};
+function isBidHex (bid) {
+  const hexRegex = /^[0-9a-fA-F]+$/;
+  return hexRegex.test(bid);
+}
 
-export async function streamLiveRender(streamInfo, res) {
+function validateFilePath(filePath) {
+  const basePath = path.resolve('./tmp'); // Set your base directory
+  const resolvedPath = path.resolve(basePath, filePath);
+
+  // Check if the resolved path is still within the base directory
+  if (resolvedPath.startsWith(basePath)) {
+    return true;
+  } else {
+    return false;
+  }
+}
+
+export async function poolStream(streamInfo, res, req) {
+  const bid = streamInfo.bid;
+
+  if (!isBidHex(bid)) {
+    res.status(403).json({ error: "Forbidden bid." });
+    return;
+  }
+
+  const removeFile = (path) => {
+    unlink(path, (error) => {
+      if (error) {
+        console.log(error);
+      }
+    });
+  }
+
+  const outputPath = `./tmp/${bid}.mp4`;
+
+  if (!validateFilePath(outputPath)) {
+    res.status(403).json({ error: "Wrong path." });
+    return;
+  }
+
+  req.on('close', () => {
+    if (!res.headersSent) {
+      const process = ongoingSpawn.get(bid);
+      killProcess(process);
+      removeFile(outputPath);
+    }
+  })
+  
+  try {
+    const inter = setInterval(() => {
+      const process = ongoingProcesses.get(bid);
+      if (process.status === 'finished') {
+        access(outputPath, constants.F_OK, (err) => {
+          if (err) {
+            console.log('File does not exist.');
+          } else {
+            const rStream = createReadStream(outputPath);
+  
+            req.on('close', () => {
+              console.log('req close');
+              rStream.unpipe(res);
+              rStream.destroy();
+              closeResponse(res);
+            })
+  
+            rStream.on('error', (err) => {
+              console.error(err);
+              res.status(500).json({ error: 'Error converting.' });
+              removeFile(outputPath);
+            });
+      
+            res.setHeader('Content-Disposition', contentDisposition(streamInfo.filename));
+            res.setHeader('Content-Type', 'video/mp4');
+      
+            rStream.on('end', () => {
+              console.log('read stream end, no more data');
+            })
+      
+            rStream.on('close', () => {
+              ongoingProcesses.delete(bid);
+              console.log('read stream close.');
+              removeFile(outputPath);
+            });
+      
+            rStream.pipe(res);
+            clearInterval(inter);
+          }
+        });
+      }
+    }, 1000);
+  
+    setTimeout(() => {
+      if (!res.headersSent) {
+        clearInterval(inter);
+        res.status(202).json({ status: 'pending', bid });
+      }
+    }, 25000);
+  } catch (err) {
+    console.log(err);
+    res.status(500).json({ error: 'Internal error' });
+  }
+}
+
+export async function streamLiveRender(streamInfo, res, req) {
+  if (!req) { return res.sendStatus(500) }
+
   let abortController = new AbortController(), process;
   const shutdown = () => (closeRequest(abortController), killProcess(process), closeResponse(res));
   const removeFile = (path) => {
     unlink(path, (error) => {
       if (error) {
-        console.error(error);
+        console.log(error)
       }
     });
   }
-
-  const outputPath = `./tmp/${(Math.random() * 99999999).toString(16)}${simpleHash(streamInfo.filename)}.mp4`;
+  const hashId = createHash('sha256').update(`${randomBytes(25).toString('base64').slice(0, 25)}${streamInfo.filename}`).digest('hex');
+  const outputPath = `./tmp/${hashId}.mp4`;
   try {
     if (streamInfo.urls.length !== 2) return shutdown();
-
     let format = streamInfo.filename.split('.')[streamInfo.filename.split('.').length - 1],
       args = [
         '-loglevel', '-8',
@@ -96,7 +222,7 @@ export async function streamLiveRender(streamInfo, res) {
 
     args = args.concat(ffmpegArgs[format]);
     if (streamInfo.metadata) args = args.concat(metadataManager(streamInfo.metadata));
-    args.push('-f', 'mp4', outputPath);
+    args.push('-f', format, outputPath);
 
     process = spawn(ffmpeg, args, {
       windowsHide: true,
@@ -105,37 +231,76 @@ export async function streamLiveRender(streamInfo, res) {
         'pipe', 'pipe'
       ],
     });
+    ongoingSpawn.set(hashId, process);
+    ongoingProcesses.set(hashId, { status: 'pending' });
+
+    req.on('close', () => {
+      if (!res.headersSent) {
+        console.log('request closed stream live');
+        shutdown();
+      }
+    })
+
+    const reqTimeout = setTimeout(() => {
+      res.status(202).json({ status: 'pending', bid: hashId });
+    }, 25000);
 
     process.on('error', (err) => {
+      ongoingProcesses.delete(hashId);  
       console.error('FFmpeg process error:', err);
       removeFile(outputPath);
     });
 
     process.on('exit', (code, signal) => {
       if (code !== 0) {
+        ongoingProcesses.delete(hashId);
         console.error(`FFmpeg process exited with code ${code} and signal ${signal}`);
+        if (!res.headersSent) {
+          res.status(500).json({ status: 'error', text: 'process stopped.' });
+        }
         removeFile(outputPath);
       } else {
+        ongoingProcesses.set(hashId, 'finished');
         console.log('FFmpeg process completed successfully');
       }
     });
-
+    
     process.on('close', () => {
-      const rStream = createReadStream(outputPath);
-      rStream.on('error', () => {
-        console.error(error);
-        shutdown();
-        res.status(500).json({ error: 'Error converting audio.' });
-        removeFile(outputPath);
-      });
+      console.log('process close');
 
-      res.setHeader('Content-Disposition', contentDisposition(streamInfo.filename));
-      res.setHeader('Content-Type', 'video/mp4');
+      clearTimeout(reqTimeout);
 
-      rStream.pipe(res);
-      rStream.on('close', () => {
-        removeFile(outputPath);
-      });
+      if (!res.headersSent) {
+        const rStream = createReadStream(outputPath);
+        rStream.on('error', () => {
+          shutdown();
+          res.status(500).json({ error: 'Error converting.' });
+          removeFile(outputPath);
+        });
+
+        req.on('close', () => {
+          rStream.unpipe(res);
+          rStream.destroy();
+          closeResponse(res);
+        })
+  
+        res.setHeader('Content-Disposition', contentDisposition(streamInfo.filename));
+        res.setHeader('Content-Type', 'video/mp4');
+  
+        rStream.on('end', () => {
+          console.log('read stream end, no more data');
+        })
+  
+        rStream.on('close', () => {
+          ongoingProcesses.delete(hashId);
+          console.log('read stream close.');
+          removeFile(outputPath);
+        });
+  
+        rStream.pipe(res);
+      } else {
+        ongoingProcesses.set(hashId, { status: 'finished' });
+      }
     });
 
   } catch (e) {
